Revert "Feat drvgo upload"
