+++ package.json
@@ -27,7 +27,6 @@
         "next": "14.2.4",
         "react": "^18",
         "react-dom": "^18",
-        "react-fast-marquee": "^1.6.5",
         "react-icons": "^5.2.1",
         "react-player": "^2.16.0",
         "react-tweet": "^3.2.1",
