+++ src/app/components/Layout.tsx
@@ -1,22 +1,20 @@
 // components/Layout.tsx
 import React, { ReactNode } from "react";
-import Footer from "./home/footer";
 import Navbar from "./home/navbar";
-import { NoticesFetch } from "./notices-fetch";
+import Footer from "./home/footer";
 
 interface LayoutProps {
-    children: ReactNode;
+  children: ReactNode;
 }
 
 const Layout: React.FC<LayoutProps> = ({ children }) => {
-    return (
-        <div>
-            <Navbar />
-            <NoticesFetch />
-            <main>{children}</main>
-            <Footer />
-        </div>
-    );
+  return (
+    <div>
+      <Navbar />
+      <main>{children}</main>
+      <Footer />
+    </div>
+  );
 };
 
 export default Layout;
