+++ src/app/components/materials/material-page-fetch.tsx
@@ -16,10 +16,37 @@ export async function MaterialPageFetch() {
     const materials: Material[] = [];
 
     for (const file of files.files) {
-        const isNotice = file.name.split("_")[0] === "imp";
-        if (isNotice) continue;
         materials.push(generateFileMetadataFromFile(file));
     }
+
+    // let categoryFolders: any[] = [];
+
+    // try {
+    //     categoryFolders = await fs.readdir(UPLOADS_FOLDER_ROUTE);
+    // } catch (error) {
+    //     categoryFolders = [];
+    // }
+
+    // const materials = [];
+
+    // for (const category of categoryFolders) {
+    //     const files = await fs.readdir(`${UPLOADS_FOLDER_ROUTE}/${category}`);
+
+    //     for (const file of files) {
+    //         const fileName = file;
+    //         const fileCreatedAt = file.split("_").pop();
+    //         const fileId = file.split("_").slice(-2, -1)[0];
+    //         const url = `/uploads/${category}/${file}`;
+
+    //         materials.push({
+    //             id: fileId,
+    //             name: fileName,
+    //             createdAt: new Date(+fileCreatedAt?.split(".")[0]!),
+    //             category,
+    //             url,
+    //         });
+    //     }
+    // }
 
     return <MaterialPage materials={materials} isAuth={isAuth} />;
 }
